backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
langchain==0.3.9
langchain-community==0.3.9
langchain-groq==0.2.1
chromadb==0.5.23
sentence-transformers==3.3.1
pydantic==2.10.3
